move releasability in its own package like that we have some space to group things together, we can explode it into multiple files and enjoy the beauty of a nice environment to work with
