+++ main/release/main.py
@@ -2,13 +2,13 @@ import os
 
 from dryable import Dryable
 from release.exceptions.invalid_input_parameters_exception import InvalidInputParametersException
+from release.releasability.releasability import Releasability
 from release.steps.ReleaseRequest import ReleaseRequest
 from release.utils.artifactory import Artifactory
 from release.utils.binaries import Binaries
 from release.utils.burgr import Burgr
 from release.utils.dryrun import DryRunHelper
 from release.utils.github import GitHub
-from release.releasability import Releasability
 from release.utils.release import publish_all_artifacts_to_binaries, releasability_checks, revoke_release, set_output
 from release.utils.slack import notify_slack
 from release.vars import binaries_bucket_name, burgrx_password, burgrx_url, burgrx_user
