+++ main/tests/releasability_check_test.py
@@ -2,9 +2,9 @@ import os
 import unittest
 from unittest.mock import patch, mock_open
 
+from release.releasability.releasability import Releasability
 from release.releasability_check import do_releasability_checks
 from release.utils.burgr import Burgr
-from release.releasability import Releasability
 
 
 class ReleasabilityCheckTest(unittest.TestCase):
